+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Fixed an issue where high resolution textures weren't read properly from legacy wismt files.
 * Fixed an issue where map textures did not always correctly load the base mip level.
-* Fixed an issue where generated JSON shader database entries had correct ordering for maps.
+* Fixed an issue where generated JSON shader database entries had incorrect ordering for maps.
 * Fixed an issue where meshes past the base level of detail (LOD) would not use correct skin weights.
 
 ### Changed