+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### unreleased
+### Changed
+* Optimized glTF file sizes by only including referenced vertex buffers.
+
 ### 0.4.1 - 2024-01-18
 ### Fixed
 * Fixed an issue where texture dimensions were reported incorrectly for xc3_wgpu.