+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### unreleased
+### 0.5.0 - 2024-01-27
 ### Fixed
 * Fixed an issue where `Msrd::from_extracted_files` would sometimes incorrectly calculate streaming data offsets.
 * Fixed an issue where some `.wilay` files would not properly rebuild all data when saving.